requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
defusedxml==0.7.1
Django==3.2.13
django-cors-headers==3.13.0
django-debug-toolbar==3.2.2
django-environ==0.4.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
environ==1.0
idna==3.3
oauthlib==3.2.0
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.9
